requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
langchain==0.1.0
torch==2.1.0+cpu
torch-geometric==2.5.0
scikit-learn==1.3.2
feedparser==6.0.10
beautifulsoup4==4.12.2
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3